let the drawsystem draw the level
