+++ dungeon/test/contrib/utils/components/item/ItemTest.java
@@ -17,7 +17,6 @@ import core.systems.LevelSystem;
 import core.utils.IVoidFunction;
 import core.utils.Point;
 import core.utils.components.draw.Animation;
-import core.utils.components.draw.Painter;
 import core.utils.components.path.SimpleIPath;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -37,7 +36,7 @@ public class ItemTest {
   /** WTF? . */
   @BeforeEach
   public void before() {
-    Game.add(new LevelSystem(Mockito.mock(Painter.class), Mockito.mock(IVoidFunction.class)));
+    Game.add(new LevelSystem(Mockito.mock(IVoidFunction.class)));
 
     TileLevel level =
         new TileLevel(
