+++ dungeon/test/core/level/utils/LevelUtilsTest.java
@@ -8,7 +8,6 @@ import core.level.TileLevel;
 import core.systems.LevelSystem;
 import core.utils.IVoidFunction;
 import core.utils.Point;
-import core.utils.components.draw.Painter;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -28,7 +27,7 @@ public class LevelUtilsTest {
   /** WTF? . */
   @BeforeEach
   public void setup() {
-    Game.add(new LevelSystem(Mockito.mock(Painter.class), Mockito.mock(IVoidFunction.class)));
+    Game.add(new LevelSystem(Mockito.mock(IVoidFunction.class)));
 
     Game.currentLevel(
         new TileLevel(
