+++ dungeon/test/core/systems/CameraSystemTest.java
@@ -40,7 +40,7 @@ public class CameraSystemTest {
     Mockito.when(startTile.position()).thenReturn(testPoint);
     Mockito.when(level.randomTilePoint(Mockito.any())).thenReturn(Optional.of(testPoint));
     Mockito.when(level.startTile()).thenReturn(startTile);
-    Game.add(new LevelSystem(null, () -> {}));
+    Game.add(new LevelSystem(() -> {}));
   }
 
   /** WTF? . */
