+++ dungeon/test/core/systems/LevelSystemTest.java
@@ -13,15 +13,9 @@ import core.level.Tile;
 import core.level.TileLevel;
 import core.level.elements.ILevel;
 import core.level.elements.tile.ExitTile;
-import core.level.utils.Coordinate;
-import core.level.utils.LevelElement;
 import core.utils.IVoidFunction;
 import core.utils.Point;
-import core.utils.components.draw.Painter;
-import core.utils.components.draw.PainterConfig;
 import core.utils.components.draw.TextureMap;
-import core.utils.components.path.IPath;
-import core.utils.components.path.SimpleIPath;
 import java.io.IOException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -34,7 +28,6 @@ import org.mockito.Mockito;
 public class LevelSystemTest {
 
   private LevelSystem api;
-  private Painter painter;
   private IVoidFunction onLevelLoader;
   private ILevel level;
 
@@ -52,10 +45,9 @@ public class LevelSystemTest {
     }
     when(textureMap.textureAt(any())).thenReturn(texture);
 
-    painter = Mockito.mock(Painter.class);
     onLevelLoader = Mockito.mock(IVoidFunction.class);
     level = Mockito.mock(TileLevel.class);
-    api = new LevelSystem(painter, onLevelLoader);
+    api = new LevelSystem(onLevelLoader);
     Game.add(api);
   }
 
@@ -75,90 +67,6 @@ public class LevelSystemTest {
     verify(onLevelLoader).execute();
     Mockito.verifyNoMoreInteractions(onLevelLoader);
     assertEquals(level, LevelSystem.level());
-  }
-
-  /** WTF? . */
-  @Test
-  public void test_execute_draw() {
-    IPath textureT1 = new SimpleIPath("dummyPath1");
-    IPath textureT2 = new SimpleIPath("dummyPath2");
-    IPath textureT3 = new SimpleIPath("dummyPath3");
-    IPath textureT4 = new SimpleIPath("dummyPath4");
-    Coordinate coordinateT1 = new Coordinate(0, 0);
-    Coordinate coordinateT2 = new Coordinate(0, 1);
-    Coordinate coordinateT3 = new Coordinate(1, 0);
-    Coordinate coordinateT4 = new Coordinate(1, 1);
-    LevelElement elementT1 = LevelElement.WALL;
-    LevelElement elementT2 = LevelElement.EXIT;
-    LevelElement elementT3 = LevelElement.WALL;
-    LevelElement elementT4 = LevelElement.SKIP;
-    Tile[][] layout = new Tile[2][2];
-    layout[0][0] = Mockito.mock(Tile.class);
-    when(layout[0][0].levelElement()).thenReturn(elementT1);
-    when(layout[0][0].visible()).thenReturn(true);
-    when(layout[0][0].tintColor()).thenReturn(-1);
-    when(layout[0][0].texturePath()).thenReturn(textureT1);
-    when(layout[0][0].position()).thenReturn(coordinateT1.toPoint());
-    layout[0][1] = Mockito.mock(Tile.class);
-    when(layout[0][1].levelElement()).thenReturn(elementT2);
-    when(layout[0][1].visible()).thenReturn(true);
-    when(layout[0][1].tintColor()).thenReturn(-1);
-
-    when(layout[0][1].texturePath()).thenReturn(textureT2);
-    when(layout[0][1].position()).thenReturn(coordinateT2.toPoint());
-    layout[1][0] = Mockito.mock(Tile.class);
-    when(layout[1][0].levelElement()).thenReturn(elementT3);
-    when(layout[1][0].visible()).thenReturn(true);
-    when(layout[1][0].tintColor()).thenReturn(-1);
-
-    when(layout[1][0].texturePath()).thenReturn(textureT3);
-    when(layout[1][0].position()).thenReturn(coordinateT3.toPoint());
-    layout[1][1] = Mockito.mock(Tile.class);
-    when(layout[1][1].levelElement()).thenReturn(elementT4);
-    when(layout[1][1].visible()).thenReturn(false);
-    when(layout[1][1].tintColor()).thenReturn(-1);
-
-    when(layout[1][1].texturePath()).thenReturn(textureT4);
-    when(layout[1][1].position()).thenReturn(coordinateT4.toPoint());
-
-    when(level.layout()).thenReturn(layout);
-
-    api.loadLevel(level);
-    api.execute();
-
-    verify(level).layout();
-    verifyNoMoreInteractions(level);
-
-    verify(layout[0][0]).levelElement();
-    verify(layout[0][0]).visible();
-    verify(layout[0][0]).texturePath();
-    verify(layout[0][0]).tintColor();
-    verify(layout[0][0]).position();
-    // for some reason mockito.verify can't compare the points of the tile correctly
-    verify(painter, times(3)).draw(any(Point.class), any(IPath.class), any(PainterConfig.class));
-    verifyNoMoreInteractions(layout[0][0]);
-
-    verify(layout[0][1]).levelElement();
-    verify(layout[0][1]).visible();
-    verify(layout[0][1]).texturePath();
-    verify(layout[0][1]).tintColor();
-    verify(layout[0][1]).position();
-    // for some reason mockito.verify can't compare the points of the tile correctly
-    verify(painter, times(3)).draw(any(Point.class), any(IPath.class), any(PainterConfig.class));
-    verifyNoMoreInteractions(layout[0][1]);
-    verify(layout[1][0]).levelElement();
-    verify(layout[1][0]).visible();
-    verify(layout[1][0]).texturePath();
-    verify(layout[1][0]).tintColor();
-    verify(layout[1][0]).position();
-    // for some reason mockito.verify can't compare the points of the tile correctly
-    verify(painter, times(3)).draw(any(Point.class), any(IPath.class), any(PainterConfig.class));
-    verifyNoMoreInteractions(layout[1][0]);
-
-    // do not draw skip tiles
-    verify(layout[1][1]).levelElement();
-    verifyNoMoreInteractions(layout[1][1]);
-    verifyNoMoreInteractions(painter);
   }
 
   /** WTF? . */
